Return all nodes that match any interface
